test(discover): add tests for the three versioning models in multi-module detection
