+++ internal/commands/discover/workflow_config_test.go
@@ -1,0 +1,259 @@
+package discover
+
+import (
+	"os"
+	"strings"
+	"testing"
+
+	"github.com/indaco/sley/internal/config"
+	"github.com/indaco/sley/internal/discovery"
+	"github.com/indaco/sley/internal/parser"
+)
+
+func TestGenerateConfigYAML_DefaultsOnly(t *testing.T) {
+	plugins := []string{"commit-parser", "tag-manager"}
+	data, err := generateConfigYAML(".version", plugins, nil)
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Check header
+	if !contains(content, "# sley configuration file") {
+		t.Error("missing header comment")
+	}
+	if !contains(content, "# Generated by 'sley discover'") {
+		t.Error("missing generated by comment")
+	}
+
+	// Check plugins
+	if !contains(content, "commit-parser: true") {
+		t.Error("missing commit-parser config")
+	}
+	if !contains(content, "tag-manager:") {
+		t.Error("missing tag-manager config")
+	}
+}
+
+func TestGenerateConfigYAML_WithDependencyCheck(t *testing.T) {
+	plugins := []string{"commit-parser", "tag-manager", "dependency-check"}
+	candidates := []discovery.SyncCandidate{
+		{
+			Path:   "package.json",
+			Format: parser.FormatJSON,
+			Field:  "version",
+		},
+	}
+
+	data, err := generateConfigYAML(".version", plugins, candidates)
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Check dependency-check config
+	if !contains(content, "dependency-check:") {
+		t.Error("missing dependency-check config")
+	}
+	if !contains(content, "auto-sync: true") {
+		t.Error("missing auto-sync config")
+	}
+	if !contains(content, "path: package.json") {
+		t.Error("missing file path in config")
+	}
+}
+
+func TestGenerateConfigYAML_EmptyPlugins(t *testing.T) {
+	data, err := generateConfigYAML(".version", []string{}, nil)
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Should still have header
+	if !contains(content, "# sley configuration file") {
+		t.Error("missing header comment")
+	}
+
+	// Should have path
+	if !contains(content, "path: .version") {
+		t.Error("missing path in config")
+	}
+}
+
+func TestGenerateConfigYAML_AllPluginTypes(t *testing.T) {
+	plugins := []string{"commit-parser", "tag-manager", "dependency-check"}
+	candidates := []discovery.SyncCandidate{
+		{Path: "package.json", Format: parser.FormatJSON, Field: "version"},
+	}
+
+	data, err := generateConfigYAML(".version", plugins, candidates)
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Check all plugins are present
+	if !contains(content, "commit-parser: true") {
+		t.Error("missing commit-parser")
+	}
+	if !contains(content, "tag-manager:") {
+		t.Error("missing tag-manager")
+	}
+	if !contains(content, "dependency-check:") {
+		t.Error("missing dependency-check")
+	}
+}
+
+func TestGenerateConfigYAMLWithWorkspace(t *testing.T) {
+	plugins := []string{"commit-parser", "tag-manager"}
+	candidates := []discovery.SyncCandidate{
+		{Path: "package.json", Format: parser.FormatJSON, Field: "version"},
+	}
+
+	data, err := generateConfigYAMLWithWorkspace(".version", plugins, candidates)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Check header comments
+	if !contains(content, "# sley configuration file") {
+		t.Error("missing header comment")
+	}
+	if !contains(content, "workspace configuration") {
+		t.Error("missing workspace comment")
+	}
+
+	// Check workspace config
+	if !contains(content, "workspace:") {
+		t.Error("missing workspace section")
+	}
+	if !contains(content, "discovery:") {
+		t.Error("missing discovery section")
+	}
+	if !contains(content, "enabled: true") {
+		t.Error("missing enabled: true")
+	}
+	if !contains(content, "recursive: true") {
+		t.Error("missing recursive: true")
+	}
+	if !contains(content, "module_max_depth: 10") {
+		t.Error("missing module_max_depth")
+	}
+	if !contains(content, "testdata") {
+		t.Error("missing testdata in exclude")
+	}
+
+	// Check plugins
+	if !contains(content, "commit-parser: true") {
+		t.Error("missing commit-parser")
+	}
+	if !contains(content, "tag-manager:") {
+		t.Error("missing tag-manager")
+	}
+}
+
+func TestMarshalConfigWithComments(t *testing.T) {
+	cfg := &config.Config{
+		Path: ".version",
+	}
+
+	data, err := marshalConfigWithComments(cfg, []string{"commit-parser"})
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Check header comments
+	if !contains(content, "# sley configuration file") {
+		t.Error("missing header comment")
+	}
+	if !contains(content, "# Generated by 'sley discover'") {
+		t.Error("missing generation comment")
+	}
+
+	// Check plugin list
+	if !contains(content, "#   - commit-parser") {
+		t.Error("missing plugin list comment")
+	}
+}
+
+func TestMarshalConfigWithComments_EmptyPlugins(t *testing.T) {
+	cfg := &config.Config{
+		Path: ".version",
+	}
+
+	data, err := marshalConfigWithComments(cfg, []string{})
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+
+	// Should have header but no plugin list
+	if !contains(content, "# sley configuration file") {
+		t.Error("missing header")
+	}
+}
+
+func TestMarshalToYAML(t *testing.T) {
+	cfg := &config.Config{
+		Path: ".version",
+	}
+
+	data, err := marshalToYAML(cfg)
+
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	content := string(data)
+	if !strings.Contains(content, "path: .version") {
+		t.Errorf("expected path in YAML, got: %q", content)
+	}
+}
+
+func TestDefaultVersionPath(t *testing.T) {
+	path := defaultVersionPath()
+	if path != ".version" {
+		t.Errorf("defaultVersionPath() = %q, want %q", path, ".version")
+	}
+}
+
+func TestConfigExists(t *testing.T) {
+	t.Run("config exists", func(t *testing.T) {
+		tmpDir := t.TempDir()
+		t.Chdir(tmpDir)
+
+		// Create .sley.yaml
+		if err := os.WriteFile(".sley.yaml", []byte("path: .version\n"), 0644); err != nil {
+			t.Fatal(err)
+		}
+
+		if !configExists() {
+			t.Error("expected configExists() to return true")
+		}
+	})
+
+	t.Run("config does not exist", func(t *testing.T) {
+		tmpDir := t.TempDir()
+		t.Chdir(tmpDir)
+
+		if configExists() {
+			t.Error("expected configExists() to return false")
+		}
+	})
+}
